# packages/api-clients/src/libs/patient-api/models/Patient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Patient Management API
 * API for managing patient profiles
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PatientContactInfo,
    PatientContactInfoFromJSON,
    PatientContactInfoFromJSONTyped,
    PatientContactInfoToJSON,
} from './';

/**
 * 
 * @export
 * @interface Patient
 */
export interface Patient {
    /**
     * Unique identifier for the patient
     * @type {string}
     * @memberof Patient
     */
    id: string;
    /**
     * Patient's full name
     * @type {string}
     * @memberof Patient
     */
    fullName: string;
    /**
     * Patient's date of birth
     * @type {Date}
     * @memberof Patient
     */
    dateOfBirth: Date;
    /**
     * 
     * @type {PatientContactInfo}
     * @memberof Patient
     */
    contactInfo: PatientContactInfo;
    /**
     * ADHD diagnosis status
     * @type {string}
     * @memberof Patient
     */
    adhdDiagnosis: PatientAdhdDiagnosis;
    /**
     * 
     * @type {string}
     * @memberof Patient
     */
    additionalNotes?: string | null;
}

/**
* @export
* @enum {string}
*/
export enum PatientAdhdDiagnosis {
    Mild = 'Mild',
    Moderate = 'Moderate',
    Severe = 'Severe'
}

export function PatientFromJSON(json: any): Patient {
    return PatientFromJSONTyped(json, false);
}

export function PatientFromJSONTyped(json: any, ignoreDiscriminator: boolean): Patient {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'fullName': json['fullName'],
        'dateOfBirth': (new Date(json['dateOfBirth'])),
        'contactInfo': PatientContactInfoFromJSON(json['contactInfo']),
        'adhdDiagnosis': json['adhdDiagnosis'],
        'additionalNotes': !exists(json, 'additionalNotes') ? undefined : json['additionalNotes'],
    };
}

export function PatientToJSON(value?: Patient | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'fullName': value.fullName,
        'dateOfBirth': (value.dateOfBirth.toISOString().substr(0,10)),
        'contactInfo': PatientContactInfoToJSON(value.contactInfo),
        'adhdDiagnosis': value.adhdDiagnosis,
        'additionalNotes': value.additionalNotes,
    };
}